+++ CJBItemSpawner/ModEntry.cs
@@ -22,6 +22,9 @@ namespace CJBItemSpawner
         /// <summary>The internal mod data about items.</summary>
         private ModItemData ItemData;
 
+        /// <summary>The item category filters available in the item spawner menu.</summary>
+        public ModDataCategory[] Categories { get; set; }
+
 
         /*********
         ** Public methods
@@ -34,15 +37,18 @@ namespace CJBItemSpawner
             this.Config = helper.ReadConfig<ModConfig>();
             this.Monitor.Log($"Started with menu key {this.Config.ShowMenuKey}.");
 
-            // read data
+            // read item data
             this.ItemData = helper.Data.ReadJsonFile<ModItemData>("assets/item-data.json");
             if (this.ItemData?.ProblematicItems == null)
                 this.Monitor.Log("One of the mod files (assets/item-data.json) is missing or invalid. Some features may not work correctly; consider reinstalling the mod.", LogLevel.Warn);
 
-            // init translations
-            I18n.Init(helper.Translation);
+            // read categories
+            this.Categories = helper.Data.ReadJsonFile<ModDataCategory[]>("assets/categories.json");
+            if (this.Categories == null)
+                this.Monitor.LogOnce("One of the mod files (assets/categories.json) is missing or invalid. Some features may not work correctly; consider reinstalling the mod.", LogLevel.Warn);
 
-            // hook events
+            // init mod
+            I18n.Init(helper.Translation);
             helper.Events.Input.ButtonPressed += this.OnButtonPressed;
         }
 
@@ -59,11 +65,27 @@ namespace CJBItemSpawner
                 return;
 
             if (e.Button == this.Config.ShowMenuKey)
-                Game1.activeClickableMenu = new ItemMenu(this.GetSpawnableItems());
+                Game1.activeClickableMenu = this.BuildMenu();
+        }
+
+        /// <summary>Build an item spawner menu.</summary>
+        private ItemMenu BuildMenu()
+        {
+            SpawnableItem[] items = this.GetSpawnableItems().ToArray();
+            string[] categoryLabels = this.GetCategories().ToArray();
+
+            return new ItemMenu(items, categoryLabels);
+        }
+
+        /// <summary>Get the available filter category labels.</summary>
+        private IEnumerable<string> GetCategories()
+        {
+            return this.Categories?.Select(p => (string)I18n.GetByKey(p.Label).Default(p.Label))
+                ?? new string[0];
         }
 
         /// <summary>Get the items which can be spawned.</summary>
-        private SearchableItem[] GetSpawnableItems()
+        private IEnumerable<SpawnableItem> GetSpawnableItems()
         {
             IEnumerable<SearchableItem> items = new ItemRepository().GetAll();
 
@@ -76,7 +98,16 @@ namespace CJBItemSpawner
                 items = items.Where(item => !problematicItems.Contains(Tuple.Create(item.Type, item.ID)));
             }
 
-            return items.ToArray();
+            // yield models
+            foreach (SearchableItem entry in items)
+            {
+                ModDataCategory category = this.Categories?.FirstOrDefault(rule => rule.IsMatch(entry.Item));
+                string categoryLabel = category != null
+                    ? I18n.GetByKey(category.Label).Default(category.Label)
+                    : I18n.Tabs_Miscellaneous();
+
+                yield return new SpawnableItem(entry, categoryLabel);
+            }
         }
     }
 }