+++ CJBItemSpawner/release-notes.md
@@ -5,6 +5,7 @@
 * Redesigned menu to make it simpler, more intuitive, and more mobile-friendly.
 * Added full controller support.
 * Added shirts normally only available during character customization.
+* Added support for customizing the filter categories via `assets/categories.json`.
 * Internal refactoring to simplify future game updates.
 
 ## 1.12.2